BUG/MINOR: cfgparse: Use the right proxy list during validity check of sinks
The commit 7af321580 ("BUG/MEDIUM: sink: bad init sequence on tcp sink from
a ring.") was not properly backported. The main loop on the proxy list in
check_config_validity() is not initialized with the right list when validity
checks are performed on sinks. It leads to errors on startup when a ring is
configured.

It is a 2.2-specific commit. Thus there is no upstream ID. No backport
needed.