BUG/MINOR: hlua: Fix Channel.line and Channel.data behavior regarding the doc
These both functions are buggy and don't respect the documentation. They
must wait for more data, if possible.

For Channel.data(), it must happen if not enough data was received orf if no
length was specified and no data was received. The first case is properly
handled but not the second one. An empty string is return instead. In
addition, if there is no data and the channel can't receive more data, 'nil'
value must be returned.

In the same spirit, for Channel.line(), we must try to wait for more data
when no line is found if not enough data was received or if no length was
specified. Here again, only the first case is properly handled. And for this
function too, 'nil' value must be returned if there is no data and the
channel can't receive more data.

This patch is related to the issue #1993. It must be backported as far as
2.5.

(cherry picked from commit 0ae2e63d85d1da9c20b2a8b511cd8ced8558e769)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 08554a4669b92f4d6c988a7dd4fd318c9e8ec903)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>