BUG/MINOR: fcgi-app: prevent 'use-fcgi-app' in default section
Despite the doc saying that 'use-fcgi-app' keyword may only be used in backend
or listen section, we forgot to prevent its usage in default section.

This is wrong because fcgi relies on a filter, and filters cannot be defined in
a default section.

Making sure such usage reports an error to the user and complies with the doc.

This could be backported up to 2.2.

(cherry picked from commit d49a580fdaebbd185a9ee6ede072fc13bec5d5bf)
Signed-off-by: Willy Tarreau <w@1wt.eu>