MINOR: listener: remove the now useless LI_F_QUIC_LISTENER flag
This flag is only used to tag a QUIC listener, which we now know by
its bind_conf's xprt as well. It's only used to decide whether or not
to perform an extra initialization step on the listener. Let's drop it
as well as the flags field.

With the various fields and options moved, the listener struct reduced
by 48 bytes total.

(cherry picked from commit 9e2682afed98aaacfb2afda8ffccbeffc0d8ce91)
 [ad: This backport is necessary to fix github issue #2141. Without it,
  every QUIC request cause a segfault.
  Backport note:
  - flags field is kept for listener struct as this is reused for
    other purposes outside of QUIC.
  - quic_add_listener is already removed in 2.7]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>