BUG/MINOR: promex: Don't forget to consume the request on error
When the promex applet triggers an error, for instance because the URI is
invalid, we must still take care to consume the request. Otherwise, the
error will be handled by HTTP analyzers as a server abort.

This patch must be backported as far as 2.0.

(cherry picked from commit c1b013bc61e165cd54fc3e4c538f95ba0c0104c6)
Signed-off-by: Willy Tarreau <w@1wt.eu>