BUG/MINOR: http-ana: Report SF_FINST_R flag on error waiting the request body
When we wait for the request body, we are still in the request analysis. So
a SF_FINST_R flag must be reported in logs. Even if some data are already
received, at this staged, nothing is sent to the server.

This patch could be backported in all stable versions.

(cherry picked from commit f4569bbcc186837cc267f751635443ad591cb817)
Signed-off-by: Willy Tarreau <w@1wt.eu>