BUG/MINOR: bwlim: Fix parameters check for set-bandwidth-limit actions
First, the inspect-delay is now tested if the action is used on a
tcp-response content rule. Then, when an expressions scope is checked, we
now take care to detect the right scope depending on the ruleset used
(tcp-request, tcp-response, http-request or http-response).

This patch could be backported to 2.7.

(cherry picked from commit 6bf86c73bae90bd6f0ed97e0df7c56199327a3bf)
Signed-off-by: Willy Tarreau <w@1wt.eu>