MINOR: listener: also support "quic+" as an address prefix
While we do support quic4@ and quic6@ for listening addresses, it was
not possible to specify that we want to use an FD inherited from the
parent with QUIC. It's just a matter of making it possible to enable
a dgram-type socket and a stream-type transport, so let's add this.

Now it becomes possible to write "quic+fd@12", "quic+ipv4@addr" etc.

(cherry picked from commit 40725a4eb0beadcdaa3c60b6f8c8fe20ab53abc3)
Signed-off-by: Willy Tarreau <w@1wt.eu>