BUG/MINOR: sink: make sure to always properly unmap a file-backed ring
The munmap() call performed on exit was incorrect since it used to apply
to the buffer instead of the area, so neither the pointer nor the size
were page-aligned. This patches corrects this and also adds a call to
msync() since munmap() alone doesn't guarantee that data will be dumped.

This should be backported to 2.6.

(cherry picked from commit fb9a4765b74cc18f25a6b732a66aae952262a7e0)
Signed-off-by: Willy Tarreau <w@1wt.eu>