DEV: haring: add a new option "-r" to automatically repair broken files
In case a file-backed ring was not properly synced before being dumped,
the output can look bogus due to the head pointer not being perfectly
up to date. In this case, passing "-r" will make haring automatically
skip entries not starting with a zero, and resynchronize with the rest
of the messages.

This should be backported to 2.6.

(cherry picked from commit e06ba9031843b6a55df486f5115b6675d71e22f7)
Signed-off-by: Willy Tarreau <w@1wt.eu>